# test: Run tests faster with `nextest`