--- a/.github/workflows/appservice.yml
+++ b/.github/workflows/appservice.yml
@@ -39,6 +39,9 @@ jobs:
       - name: Load cache
         uses: Swatinem/rust-cache@v1
 
+      - name: Install nextest
+        uses: taiki-e/install-action@nextest
+
       - name: Run checks
         uses: actions-rs/cargo@v1
         with: