--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -127,6 +127,12 @@ jobs:
           command: nextest
           args: run
 
+      - name: Test documentation
+        uses: actions-rs/cargo@v1
+        with:
+          command: test
+          args: --doc
+
   test-nodejs:
     name: linux / node.js (${{ matrix.node-version }})
     if: github.event_name == 'push' || !github.event.pull_request.draft