--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -76,7 +76,10 @@ jobs:
     - name: Load cache
       uses: Swatinem/rust-cache@v1
 
-    - name: Clippy
+    - name: Install nextest
+      uses: taiki-e/install-action@nextest
+
+    - name: Test
       uses: actions-rs/cargo@v1
       with:
         command: run