--- a/xtask/src/ci.rs
+++ b/xtask/src/ci.rs
@@ -22,8 +22,6 @@ enum CiCommand {
     Clippy,
     /// Check documentation
     Docs,
-    /// Run default tests
-    Test,
     /// Run tests with a specific feature set
     TestFeatures {
         #[clap(subcommand)]
@@ -84,7 +82,6 @@ impl CiArgs {
                 CiCommand::Typos => check_typos(),
                 CiCommand::Clippy => check_clippy(),
                 CiCommand::Docs => check_docs(),
-                CiCommand::Test => run_tests(),
                 CiCommand::TestFeatures { cmd } => run_feature_tests(cmd),
                 CiCommand::TestAppservice => run_appservice_tests(),
                 CiCommand::Wasm { cmd } => run_wasm_checks(cmd),
@@ -96,7 +93,6 @@ impl CiArgs {
                 check_clippy()?;
                 check_typos()?;
                 check_docs()?;
-                run_tests()?;
                 run_feature_tests(None)?;
                 run_appservice_tests()?;
                 run_wasm_checks(None)?;
@@ -139,12 +135,6 @@ fn check_clippy() -> Result<()> {
 
 fn check_docs() -> Result<()> {
     build_docs([], DenyWarnings::Yes)
-}
-
-fn run_tests() -> Result<()> {
-    cmd!("rustup run stable cargo test").run()?;
-    cmd!("rustup run beta cargo test").run()?;
-    Ok(())
 }
 
 fn run_feature_tests(cmd: Option<FeatureSet>) -> Result<()> {