--- a/xtask/src/ci.rs
+++ b/xtask/src/ci.rs
@@ -156,6 +156,9 @@ fn run_feature_tests(cmd: Option<FeatureSet>) -> Result<()> {
     let run = |arg_set: &str| {
         cmd!("rustup run stable cargo nextest run -p matrix-sdk")
             .args(arg_set.split_whitespace())
+            .run()?;
+        cmd!("rustup run stable cargo test --doc -p matrix-sdk")
+            .args(arg_set.split_whitespace())
             .run()
     };
 
@@ -179,7 +182,9 @@ fn run_crypto_tests() -> Result<()> {
     )
     .run()?;
     cmd!("rustup run stable cargo nextest run -p matrix-sdk-crypto --features=backups_v1").run()?;
+    cmd!("rustup run stable cargo test --doc -p matrix-sdk-crypto --features=backups_v1").run()?;
     cmd!("rustup run stable cargo nextest run -p matrix-sdk-crypto-ffi").run()?;
+    cmd!("rustup run stable cargo test --doc -p matrix-sdk-crypto-ffi").run()?;
 
     Ok(())
 }
@@ -187,6 +192,7 @@ fn run_crypto_tests() -> Result<()> {
 fn run_appservice_tests() -> Result<()> {
     cmd!("rustup run stable cargo clippy -p matrix-sdk-appservice -- -D warnings").run()?;
     cmd!("rustup run stable cargo nextest run -p matrix-sdk-appservice").run()?;
+    cmd!("rustup run stable cargo test --doc -p matrix-sdk-appservice").run()?;
 
     Ok(())
 }