@Hywan 2022-06-20
With `xargs -P`, we can achieve the following results:

```
Benchmark 1: cargo test --workspace
  Time (mean ± σ):     51.785 s ±  2.066 s    [User: 183.471 s, System: 10.563 s]
  Range (min … max):   49.151 s … 56.641 s    10 runs

Benchmark 2: cargo nextest run --workspace && cargo test --doc
  Time (mean ± σ):     44.556 s ±  0.894 s    [User: 192.213 s, System: 11.441 s]
  Range (min … max):   43.170 s … 45.762 s    10 runs

Benchmark 3: echo "cargo nextest run --workspace\0cargo test --doc" | xargs -0 -P 0 -I@ sh -c @
  Time (mean ± σ):     34.339 s ±  4.357 s    [User: 192.472 s, System: 11.739 s]
  Range (min … max):   22.009 s … 36.422 s    10 runs
```

Benchmark 2 is 1.16 times faster than Benchmark 1
Benchmark 3 is 1.51 times faster than Benchmark 1

I'm not sure how it is possible to run such `xargs` command in a YAML file while having cross-compatibility, but let's try!