@Hywan 2022-06-20
Without `xargs -P`, here are the results for the “test” step of the CI workflow for some jobs (I won't do it exhaustively, time differs, it's just to provide an order of magnitude idea):

| jobs | before | after | speed up factor |
|--|--|--|--|
| linux / stable | 11m 21s | 9m 13s | **1.23** |
| macOS / stable | 15m 25s | 10m 40s | **1.45** |
| linux / features-sso-login | 9m 30s | 5m 9s | **1.84** |
| linux / features-rustls-tls | 5m 4s | 2m 13s | **2.29** |
| linux / features-no-encryption | 6m 16s | 2m 14s | **2.81** |
| linux / features-no-sled | 6m 51s | 4m 9s | **1.65** |
| linux / features-no-encryption-and-sled | 5m 10s | 1m 51s | **2.79** |

Just on those examples, running tests is 2.01 times faster on average.