@Hywan 2022-06-20
When running `cargo nextest run` and `cargo test --doc` in parallel, we have worst results on Github Actions, probably because the number of threads is limited. In that case, they are spending more time waiting on each other than truly running what we want. Let's keep things simple and let's use `cargo-nextest` for now.