@Hywan 2022-06-22
> Removal of dead code is still not a separate commit afaict but otherwise looks good.

Done, https://github.com/matrix-org/matrix-rust-sdk/pull/778/commits/d9475c131a43c1b08657804a1ec365492e7b6926 :-).